+++ docs/package.json
@@ -11,7 +11,8 @@
     "start": "next start",
     "lint": "eslint ./{components,pages,utils}",
     "lint:fix": "eslint ./{components,pages,utils} --fix",
-    "build:sitemap": "next-sitemap --config next-sitemap.config.cjs"
+    "build:sitemap": "next-sitemap --config next-sitemap.config.cjs",
+    "clean": "rm -rf .next build pages/reference/*.mdx pages/reference/**/*"
   },
   "repository": {
     "type": "git",
@@ -50,7 +51,7 @@
     "shiki": "^1.22.0",
     "tailwindcss": "^3.4.13",
     "typedoc": "^0.27.6",
-    "typedoc-plugin-markdown": "4.4.1",
+    "typedoc-plugin-markdown": "4.3.3",
     "typedoc-plugin-mdn-links": "4.0.11",
     "typedoc-plugin-no-inherit": "^1.5.0"
   }
