+++ docs/typedoc.config.cjs
@@ -32,7 +32,6 @@ module.exports = {
   excludeNotDocumented: true,
   excludeExternals: true,
   excludeInternal: true,
-  excludePrivate: true,
   excludeProtected: true,
   excludeReferences: true,
   cleanOutputDir: false,
@@ -77,8 +76,5 @@ module.exports = {
   expandObjects: true,
   parametersFormat: "table",
   indexFormat: "table",
-  textContentMappings: {
-    "label.packages": "Integrations",
-  },
   useCodeBlocks: true,
 }
