+++ packages/core/src/lib/utils/logger.ts
@@ -1,12 +1,6 @@
 import { AuthError } from "../../errors.js"
+import type { WarningCode } from "../../warnings.js"
 import type { AuthConfig } from "../../index.js"
-
-export type WarningCode =
-  | "debug-enabled"
-  | "csrf-disabled"
-  | "experimental-webauthn"
-  | "env-url-basepath-redundant"
-  | "env-url-basepath-mismatch"
 
 /**
  * Override any of the methods, and the rest will use the default logger.
@@ -47,7 +41,7 @@ const defaultLogger: LoggerInstance = {
     }
   },
   warn(code) {
-    const url = `https://warnings.authjs.dev#${code}`
+    const url = `https://warnings.authjs.dev`
     console.warn(`${yellow}[auth][warn][${code}]${reset}`, `Read more: ${url}`)
   },
   debug(message, metadata) {
