+++ packages/core/src/types.ts
@@ -58,6 +58,7 @@ import { AuthConfig } from "./index.js"
 import type { JWTOptions } from "./jwt.js"
 import type { Cookie } from "./lib/utils/cookie.js"
 import type { LoggerInstance } from "./lib/utils/logger.js"
+import type { WarningCode } from "./warnings.js"
 import type {
   CredentialsConfig,
   EmailConfig,
@@ -72,7 +73,7 @@ import type {
 
 export type { WebAuthnOptionsResponseBody } from "./lib/utils/webauthn-utils.js"
 export type { AuthConfig } from "./index.js"
-export type { LoggerInstance }
+export type { LoggerInstance, WarningCode }
 export type Awaitable<T> = T | PromiseLike<T>
 export type Awaited<T> = T extends Promise<infer U> ? U : T
